requirements.txt
asgiref==3.3.1
Django==3.1.3
django-tastypie==0.14.3
python-dateutil==2.8.1
python-mimeparse==1.6.0
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
